Redownload of failed MMS on delete
On failed MMS download, mNotificationInd is saved.
On MMS deletion request, if the mNotificationInd exists, it is handled
again.
The nuntium-inject-push command gets a new flag --denial-count N (short -dN),
which denies the test MMS servings for N times.

This is a very dirty code, with additional comments. Non essential
changes will be trimmed in the future.